Created an announcer for logging ETOS state
